+++ packages/appsync-modelgen-plugin/src/__tests__/visitors/appsync-visitor.test.ts
@@ -411,6 +411,26 @@ describe('AppSyncModelVisitor', () => {
         },
       });
     });
+
+    it('processes index with queryField', () => {
+      const schema = /* GraphQL */ `
+        type Project @model {
+          id: ID!
+          name: String @index(name: "nameIndex", queryField: "myQuery")
+        }
+      `;
+      const visitor = createAndGenerateVisitor(schema, true);
+      visitor.generate();
+      const projectKeyDirective = visitor.models.Project.directives.find(directive => directive.name === 'key');
+      expect(projectKeyDirective).toEqual({
+        name: 'key',
+        arguments: {
+          name: 'nameIndex',
+          queryField: 'myQuery',
+          fields: ['name'],
+        },
+      });
+    });
   });
 
   describe('auth directive', () => {
@@ -833,10 +853,10 @@ describe('AppSyncModelVisitor', () => {
       expect(visitor.models.Human.fields[2].directives[0].arguments.fields[0]).toEqual('governmentID');
       expect(visitor.models.Human.fields[2].directives[0].arguments.indexName).toEqual('byHuman');
       expect(visitor.models.PetFriend).toBeDefined();
-      expect(visitor.models.PetFriend.fields.length).toEqual(5);
-      expect(visitor.models.PetFriend.fields[2].directives[0].name).toEqual('belongsTo');
-      expect(visitor.models.PetFriend.fields[2].directives[0].arguments.fields.length).toEqual(1);
-      expect(visitor.models.PetFriend.fields[2].directives[0].arguments.fields[0]).toEqual('animalID');
+      expect(visitor.models.PetFriend.fields.length).toEqual(7);
+      expect(visitor.models.PetFriend.fields[4].directives[0].name).toEqual('belongsTo');
+      expect(visitor.models.PetFriend.fields[4].directives[0].arguments.fields.length).toEqual(1);
+      expect(visitor.models.PetFriend.fields[4].directives[0].arguments.fields[0]).toEqual('animalID');
       expect(visitor.models.Animal.fields.length).toEqual(5);
       expect(visitor.models.Animal.fields[2].type).toEqual('PetFriend');
       expect(visitor.models.Animal.fields[2].directives.length).toEqual(1);
@@ -856,7 +876,7 @@ describe('AppSyncModelVisitor', () => {
       expect(visitor.models.ModelA.fields[1].directives[0].arguments.indexName).toEqual('byModelA');
 
       expect(visitor.models.Models).toBeDefined();
-      expect(visitor.models.Models.fields.length).toEqual(5);
+      expect(visitor.models.Models.fields.length).toEqual(7);
 
       const modelA = visitor.models.Models.fields.find(f => f.name === 'modelA');
       expect(modelA).toBeDefined();